+++ src/main/java/com/jobnote/domain/applicationform/controller/ApplicationFormController.java
@@ -1,4 +1,4 @@
-package com.jobnote.domain.applicationform.api;
+package com.jobnote.domain.applicationform.controller;
 
 import com.jobnote.auth.config.LoginUser;
 import com.jobnote.auth.dto.CustomPrincipal;
