+++ src/main/java/com/jobnote/domain/schedule/controller/ScheduleController.java
@@ -1,4 +1,4 @@
-package com.jobnote.domain.schedule.api;
+package com.jobnote.domain.schedule.controller;
 
 import com.jobnote.auth.config.LoginUser;
 import com.jobnote.auth.dto.CustomPrincipal;
